Work on implementing the package scoped variable convertions.
For now, we just recognize some of those variable references, the
intermediate format (ir1) that the parser outputs is lacking details:
more normalization needs to happen for expression so that the walk-apply
function is able to walk through them directly.

When we have that ability, it's a matter of applying the conversion
function to the node containing the qname reference so that we can setf
its replacement to it.

The replacement might be a funcall to a plconvert provided function that
we will have to generate from the package specs and introduce in the
output for the package.