Clean-up converter code, work on out+return call sites.
And add support for properly detecting out+return functions and matching
them from the pl-funcall call sites. The implementation here is a 99%
one in that we only match function name and arity rather than resolve
the full argument list types from the call-site (which is quite a very
hard problem to solve).

It is considered so rare to have name/arity clashes on PL functions that
we can live with outputing a WARNING in those cases rather than doing
the full call site rewrite.

The call site rewriting itself isn't part of this patch, we only go as
far as allowing the next patch to decide where to issue the new function
call "protocol" by means of the parse tree node traveral list.