Out+return callsites: add support for collisions.
It could be that the same signature (name/arity) is overloaded with
normal functions and functions with out parameters and a return value.

We now recognize those cases, and when possible we match against the
actual function that is called, by determining the data types of
run-time arguments given at the call-site.

In cases where that's not possible, the intend is to add a WARNING to
the translated code. In cases where we could match which function is
actually called and that function is an out+return, the intend is to go
edit the parse tree and make the necessary change happen.